Arrumei erros entre Int e Long e erro no COnfig

+++ src/main/java/com/educandoweb/course/resources/OrderResoure.java
@@ -27,7 +27,7 @@ public class OrderResoure {
 	}
 	
 	@GetMapping(value = "/{id}")
-	public ResponseEntity<Order> findById(@PathVariable Integer id){
+	public ResponseEntity<Order> findById(@PathVariable Long id){
 		Order obj = service.findById(id);
 		return ResponseEntity.ok().body(obj);
 	}
